start moving all audio stuff into own modality file
